+++ src/meltano/core/plugin_install_service.py
@@ -210,7 +210,7 @@ class PluginInstallService:
             return state
 
         try:
-            with plugin.trigger_hooks("install", self, plugin, reason):
+            async with plugin.trigger_hooks("install", self, plugin, reason):
                 await installer_factory(self.project, plugin).install(
                     reason, self.clean
                 )