+++ tests/meltano/core/behavior/test_hookable.py
@@ -16,39 +16,40 @@ class Hooked(HookObject):
         return self._calls.copy()
 
     @hook("after_test")
-    def after_test(self):
+    async def after_test(self):
         self.call("after_test")
 
     @hook("after_test")
-    def after_test_2(self):
+    async def after_test2(self):
         self.call("after_test_2")
 
     @hook("before_test")
-    def before_test(self):
+    async def before_test(self):
         self.call("before_test")
 
     @hook("before_test")
-    def before_test_2(self):
+    async def before_test2(self):
         self.call("before_test_2")
 
 
 class DerivedHooked(Hooked):
     @hook("before_test")
-    def derived_before_test(self):
+    async def derived_before_test(self):
         super().call("derived_before_test")
 
 
 class Hooked2(HookObject):
     @hook("before_test")
-    def another_class(self):
+    async def another_class(self):
         raise Exception()
 
 
 class TestHookable:
-    def test_trigger_hook(self):
+    @pytest.mark.asyncio
+    async def test_trigger_hook(self):
         subject = Hooked()
         process = mock.MagicMock()
-        with subject.trigger_hooks("test"):
+        async with subject.trigger_hooks("test"):
             process()
 
         assert subject.calls == [
@@ -59,21 +60,24 @@ class TestHookable:
         ]
         assert process.called_once
 
-    def test_trigger_hook_raise(self):
+    @pytest.mark.asyncio
+    async def test_trigger_hook_raise(self):
         subject = Hooked()
         process = mock.MagicMock()
 
         # it raises exceptions correctly
-        with pytest.raises(Exception), subject.trigger_hooks("test"):
-            raise Exception()
+        with pytest.raises(Exception):
+            async with subject.trigger_hooks("test"):
+                raise Exception()
 
         assert subject.calls == ["before_test", "before_test_2"]
 
-    def test_trigger_derived_hook(self):
+    @pytest.mark.asyncio
+    async def test_trigger_derived_hook(self):
         subject = DerivedHooked()
 
         process = mock.MagicMock()
-        with subject.trigger_hooks("test"):
+        async with subject.trigger_hooks("test"):
             process()
 
         assert subject.calls == [